1. platform specific code refactoring: windows need to be reimplemented. it does not support SCTP natively.
